PR feedback: - cine centralized to OHIFCornerstoneViewport - introduced a type for the CinePlayer properties
