+++ extensions/cornerstone/src/Viewport/OHIFCornerstoneViewport.tsx
@@ -20,6 +20,7 @@ import {
   IVolumeViewport,
 } from '@cornerstonejs/core/dist/esm/types';
 import getSOPInstanceAttributes from '../utils/measurementServiceMappings/utils/getSOPInstanceAttributes';
+import { CinePlayer, useCine, useViewportGrid } from '@ohif/ui';
 
 const STACK = 'stack';
 
@@ -112,6 +113,9 @@ const OHIFCornerstoneViewport = React.memo(props => {
   } = props;
 
   const [scrollbarHeight, setScrollbarHeight] = useState('100px');
+  const [{ isCineEnabled, cines }, cineService] = useCine();
+  const [{ activeViewportIndex }] = useViewportGrid();
+  const [enabledVPElement, setEnabledVPElement] = useState(null);
 
   const elementRef = useRef();
 
@@ -125,6 +129,73 @@ const OHIFCornerstoneViewport = React.memo(props => {
     cornerstoneCacheService,
     viewportGridService,
   } = servicesManager.services;
+
+  const cineHandler = () => {
+    if (!cines || !cines[viewportIndex] || !enabledVPElement) {
+      return;
+    }
+
+    const cine = cines[viewportIndex];
+    const isPlaying = cine.isPlaying || false;
+    const frameRate = cine.frameRate || 24;
+
+    const validFrameRate = Math.max(frameRate, 1);
+
+    if (isPlaying) {
+      cineService.playClip(enabledVPElement, {
+        framesPerSecond: validFrameRate,
+      });
+    } else {
+      cineService.stopClip(enabledVPElement);
+    }
+  };
+
+  useEffect(() => {
+    eventTarget.addEventListener(
+      Enums.Events.STACK_VIEWPORT_NEW_STACK,
+      cineHandler
+    );
+
+    return () => {
+      cineService.setCine({ id: viewportIndex, isPlaying: false });
+      eventTarget.removeEventListener(
+        Enums.Events.STACK_VIEWPORT_NEW_STACK,
+        cineHandler
+      );
+    };
+  }, [enabledVPElement]);
+
+  useEffect(() => {
+    if (!cines || !cines[viewportIndex] || !enabledVPElement) {
+      return;
+    }
+
+    cineHandler();
+
+    return () => {
+      if (enabledVPElement && cines?.[viewportIndex]?.isPlaying) {
+        cineService.stopClip(enabledVPElement);
+      }
+    };
+  }, [cines, viewportIndex, cineService, enabledVPElement, cineHandler]);
+
+  const cine = cines[viewportIndex];
+  const isPlaying = (cine && cine.isPlaying) || false;
+
+  const handleCineClose = () => {
+    toolbarService.recordInteraction({
+      groupId: 'MoreTools',
+      itemId: 'cine',
+      interactionType: 'toggle',
+      commands: [
+        {
+          commandName: 'toggleCine',
+          commandOptions: {},
+          context: 'CORNERSTONE',
+        },
+      ],
+    });
+  };
 
   // useCallback for scroll bar height calculation
   const setImageScrollBarHeight = useCallback(() => {
@@ -176,6 +247,7 @@ const OHIFCornerstoneViewport = React.memo(props => {
       const viewportIndex = viewportInfo.getViewportIndex();
 
       setEnabledElement(viewportIndex, element);
+      setEnabledVPElement(element);
 
       const renderingEngineId = viewportInfo.getRenderingEngineId();
       const toolGroupId = viewportInfo.getToolGroupId();
@@ -360,6 +432,25 @@ const OHIFCornerstoneViewport = React.memo(props => {
         scrollbarHeight={scrollbarHeight}
         servicesManager={servicesManager}
       />
+      {isCineEnabled && (
+        <CinePlayer
+          className="absolute left-1/2 -translate-x-1/2 bottom-3"
+          isPlaying={isPlaying}
+          onClose={handleCineClose}
+          onPlayPauseChange={isPlaying =>
+            cineService.setCine({
+              id: activeViewportIndex,
+              isPlaying,
+            })
+          }
+          onFrameRateChange={frameRate =>
+            cineService.setCine({
+              id: activeViewportIndex,
+              frameRate,
+            })
+          }
+        />
+      )}
     </div>
   );
 }, areEqual);
