+++ extensions/measurement-tracking/src/viewports/TrackedCornerstoneViewport.tsx
@@ -5,17 +5,13 @@ import OHIF, { utils } from '@ohif/core';
 import {
   Notification,
   ViewportActionBar,
-  useCine,
-  useViewportGrid,
   useViewportDialog,
   Tooltip,
   Icon,
-  CinePlayer,
 } from '@ohif/ui';
 
 import { useTranslation } from 'react-i18next';
 
-import { eventTarget, Enums } from '@cornerstonejs/core';
 import { annotation } from '@cornerstonejs/tools';
 import { useTrackedMeasurements } from './../getContextModule';
 
@@ -38,19 +34,15 @@ function TrackedCornerstoneViewport(props) {
   const {
     measurementService,
     cornerstoneViewportService,
-    toolbarService,
   } = servicesManager.services;
 
   // Todo: handling more than one displaySet on the same viewport
   const displaySet = displaySets[0];
 
   const [trackedMeasurements] = useTrackedMeasurements();
-  const [{ activeViewportIndex }] = useViewportGrid();
-  const [{ isCineEnabled, cines }, cineService] = useCine();
   const [viewportDialogState] = useViewportDialog();
   const [isTracked, setIsTracked] = useState(false);
   const [trackedMeasurementUID, setTrackedMeasurementUID] = useState(null);
-  const [element, setElement] = useState(null);
 
   const { trackedSeries } = trackedMeasurements.context;
   const viewportId = viewportOptions.viewportId;
@@ -72,26 +64,6 @@ function TrackedCornerstoneViewport(props) {
     SpacingBetweenSlices,
     ManufacturerModelName,
   } = displaySet.images[0];
-
-  const cineHandler = () => {
-    if (!cines || !cines[viewportIndex] || !element) {
-      return;
-    }
-
-    const cine = cines[viewportIndex];
-    const isPlaying = cine.isPlaying || false;
-    const frameRate = cine.frameRate || 24;
-
-    const validFrameRate = Math.max(frameRate, 1);
-
-    if (isPlaying) {
-      cineService.playClip(element, {
-        framesPerSecond: validFrameRate,
-      });
-    } else {
-      cineService.stopClip(element);
-    }
-  };
 
   useEffect(() => {
     if (isTracked) {
@@ -121,49 +93,9 @@ function TrackedCornerstoneViewport(props) {
     };
   }, [isTracked]);
 
-  // unmount cleanup
-  useEffect(() => {
-    eventTarget.addEventListener(
-      Enums.Events.STACK_VIEWPORT_NEW_STACK,
-      cineHandler
-    );
-
-    return () => {
-      cineService.setCine({ id: viewportIndex, isPlaying: false });
-      eventTarget.removeEventListener(
-        Enums.Events.STACK_VIEWPORT_NEW_STACK,
-        cineHandler
-      );
-    };
-  }, [element]);
-
-  useEffect(() => {
-    if (!cines || !cines[viewportIndex] || !element) {
-      return;
-    }
-
-    cineHandler();
-
-    return () => {
-      if (element && cines?.[viewportIndex]?.isPlaying) {
-        cineService.stopClip(element);
-      }
-    };
-  }, [cines, viewportIndex, cineService, element, cineHandler]);
-
   if (trackedSeries.includes(SeriesInstanceUID) !== isTracked) {
     setIsTracked(!isTracked);
   }
-
-  /**
-   * OnElementEnabled callback which is called after the cornerstoneExtension
-   * has enabled the element. Note: we delegate all the image rendering to
-   * cornerstoneExtension, so we don't need to do anything here regarding
-   * the image rendering, element enabling etc.
-   */
-  const onElementEnabled = evt => {
-    setElement(evt.detail.element);
-  };
 
   function switchMeasurement(direction) {
     const newTrackedMeasurementUID = _getNextMeasurementUID(
@@ -190,26 +122,9 @@ function TrackedCornerstoneViewport(props) {
       '@ohif/extension-cornerstone.viewportModule.cornerstone'
     );
 
-    return <Component {...props} onElementEnabled={onElementEnabled} />;
+    return <Component {...props} />;
   };
 
-  const cine = cines[viewportIndex];
-  const isPlaying = (cine && cine.isPlaying) || false;
-
-  const handleCineClose = () => {
-    toolbarService.recordInteraction({
-      groupId: 'MoreTools',
-      itemId: 'cine',
-      interactionType: 'toggle',
-      commands: [
-        {
-          commandName: 'toggleCine',
-          commandOptions: {},
-          context: 'CORNERSTONE',
-        },
-      ],
-    });
-  };
   return (
     <>
       <ViewportActionBar
@@ -256,25 +171,6 @@ function TrackedCornerstoneViewport(props) {
             />
           )}
         </div>
-        {isCineEnabled && (
-          <CinePlayer
-            className="absolute left-1/2 -translate-x-1/2 bottom-3"
-            isPlaying={isPlaying}
-            onClose={handleCineClose}
-            onPlayPauseChange={isPlaying =>
-              cineService.setCine({
-                id: activeViewportIndex,
-                isPlaying,
-              })
-            }
-            onFrameRateChange={frameRate =>
-              cineService.setCine({
-                id: activeViewportIndex,
-                frameRate,
-              })
-            }
-          />
-        )}
       </div>
     </>
   );
